Allow Scripts To Be Written In ES6
![ ](https://i.imgflip.com/xpu5b.jpg)

ES2015
===

Lets usher in the new year by allowing those of use who write JavaScript to not
have to wince in pain anymore when we look upon the atrocity that is CoffeeScript.

Instead, lets allow our beloved developers to write with a syntax that only is
possible in that of CoffeeScripts wildest dreams. But seriously. Eff CS.

Caveats
===

Well, there are a few things I've done here that we may want to think about.

**Have I tested this**... yes.. don't worry.

However, there is something I was unable to test. That is the HipChat adapter
for hubot. There is a node engine dep of `~12.2`... I'm not sure why. I'm pretty
sure he may have just forgotten to remember that Node LTS is at 4.2.4 now. But
hey, people get busy, and OSS is OSS. So, ultimately this is probably the biggest
issue with this pull request, besides those of you who just truly hate JavaScript
and feel CoffeeScript brought you to some enlightened feeling of being able to use
this scripting language without having to actually use it. 😄

Closing
===

Anyway, enough ragging on CoffeeScript and how much it royally sucks. It would be
sweet if this works out. Otherwise, I guess we can just start writing ES5 scripts
instead.

Future Cleanup
===
Ultimately, I just want to see if this works with the hipchat adapter. If it does
there is some quick stuff we can do to separate the `.js` from the coffee script
and then just load those scripts in with an index file written in POJS or CS (idc).